overall and clean code
