+++ src/components/dashboard/farm-card.tsx
@@ -1,4 +1,3 @@
-import Image from 'next/image'
 import { Card, CardContent, CardFooter } from "@/components/ui/card"
 import { Badge } from "@/components/ui/badge"
 import { Separator } from "@/components/ui/separator"
